+++ src/main/groovy/org/myire/quill/check/AbstractCheckTaskEnhancer.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015, 2018, 2019 Peter Franzen. All rights reserved.
+ * Copyright 2015, 2018, 2019, 2021 Peter Franzen. All rights reserved.
  *
  * Licensed under the Apache License v2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
@@ -10,6 +10,7 @@ import org.gradle.api.reporting.Report
 
 import org.myire.quill.common.Projects
 import org.myire.quill.common.Tasks
+import org.myire.quill.report.Reports
 import org.myire.quill.report.ReportTransformingReport
 import org.myire.quill.report.TransformingReport
 
@@ -80,7 +81,7 @@ abstract class AbstractCheckTaskEnhancer<T extends Task>
                                                                   fTask.name.capitalize() + ' HTML report',
                                                                   pInputReport,
                                                                   pXslResource);
-        aReport.enabled = true;
+        Reports.setRequired(aReport, true);
         fTask.convention.add(TRANSFORMING_REPORT_NAME, aReport);
 
         // The XSL file used to create the transforming report is an input to the task.