+++ src/main/groovy/org/myire/quill/check/CheckstyleEnhancer.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Peter Franzen. All rights reserved.
+ * Copyright 2015, 2021 Peter Franzen. All rights reserved.
  *
  * Licensed under the Apache License v2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
@@ -13,6 +13,7 @@ import org.gradle.api.plugins.quality.CheckstyleReports
 import org.gradle.api.reporting.Report
 
 import org.myire.quill.common.Projects
+import org.myire.quill.report.Reports
 
 
 /**
@@ -27,7 +28,7 @@ class CheckstyleEnhancer extends AbstractPluginEnhancer<Checkstyle>
 {
     static private final String CHECKSTYLE_TOOL_NAME = 'Checkstyle'
     static private final String CHECKSTYLE_EXTENSION_NAME = CHECKSTYLE_TOOL_NAME.toLowerCase()
-    static private final String DEFAULT_TOOL_VERSION = '8.41.1'
+    static private final String DEFAULT_TOOL_VERSION = '8.44'
 
     // The file name to use when extracting the built-in config file from its resource.
     static private final String BUILTIN_CONFIG_FILE_NAME = 'checkstyle_config.xml'
@@ -148,8 +149,8 @@ class CheckstyleEnhancer extends AbstractPluginEnhancer<Checkstyle>
         void enhance()
         {
             // Produce an xml report but no html report.
-            task.reports.getXml().enabled = true;
-            getHtmlReport(task.reports)?.enabled = false;
+            Reports.setRequired(task.reports.getXml(), true);
+            Reports.setRequired(getHtmlReport(task.reports), false);
 
             // By default tasks get the config file from the extension, which is fine, but we must
             // check if the extension specifies the built-in config file (which it is configured to