+++ src/main/groovy/org/myire/quill/junit/JUnitSummaryReport.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Peter Franzen. All rights reserved.
+ * Copyright 2015, 2021 Peter Franzen. All rights reserved.
  *
  * Licensed under the Apache License v2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
@@ -13,6 +13,7 @@ import org.gradle.api.tasks.testing.Test
 
 import org.myire.quill.common.Projects
 import org.myire.quill.report.DefaultSingleFileReport
+import org.myire.quill.report.Reports
 
 
 /**
@@ -55,7 +56,7 @@ class JUnitSummaryReport extends DefaultSingleFileReport
         if (junitReportDirectory != null)
             return junitReportDirectory;
         else
-            return fTask.reports.getJunitXml()?.destination;
+            return Reports.getOutputLocation(fTask.reports.getJunitXml());
     }
 
 
@@ -82,11 +83,11 @@ class JUnitSummaryReport extends DefaultSingleFileReport
      */
     boolean checkUpToDate()
     {
-        if (!enabled)
+        if (!reportIsRequired())
             // A disabled report is always up-to-date.
             return true;
 
-        File aReportFile = getDestination();
+        File aReportFile = Reports.getOutputLocation(this);
         if (aReportFile == null)
             // No report file should be created and it is thereby always up-to-date.
             return true;
@@ -107,9 +108,12 @@ class JUnitSummaryReport extends DefaultSingleFileReport
      */
     void createReport()
     {
-        JUnitReportAggregator aAggregator = new JUnitReportAggregator();
-        aAggregator.aggregate(getJunitReportDirectory(), getFileNamePattern());
-        aAggregator.writeXmlFile(getDestination());
+        if (reportIsRequired())
+        {
+            JUnitReportAggregator aAggregator = new JUnitReportAggregator();
+            aAggregator.aggregate(getJunitReportDirectory(), getFileNamePattern());
+            aAggregator.writeXmlFile(Reports.getOutputLocation(this));
+        }
     }